// src/interfaces/cardholder.ts
interface ICardholder {
    country_code:       string;
    phone_number:       string;
    verification_id:    string;
    document_id:        string;
    country:            string;
    document_type:      string;
    email:              string;
    birthday:           string;
    first_name:         string;
    last_name:          string;
    street1:            string;
    street2:            string;
    locality:           string;
    region:             string;
    postal_code:        string;
}

export default ICardholder;